requirements.txt
pillow==10.4.0
pytest==8.3.2
